[REEF-1293] Fix test failures during file extraction from jar
This change:
 * makes REEF functional tests run in sequence instead of in parallel.
 * stops extracting .jar file by a test if it's already been extracted by another one.

JIRA:
  [REEF-1293](https://issues.apache.org/jira/browse/REEF-1293)

Pull request:
  This closes #923

+++ lang/cs/Org.Apache.REEF.Client/Common/DriverFolderPreparationHelper.cs
@@ -127,7 +127,10 @@ namespace Org.Apache.REEF.Client.Common
                 {
                     fileName = Path.Combine(driverFolderPath, _fileNames.GetBridgeExePath());
                 }
-                File.WriteAllBytes(fileName, resourceHelper.GetBytes(fileResources.Value));
+                if (!File.Exists(fileName))
+                {
+                    File.WriteAllBytes(fileName, resourceHelper.GetBytes(fileResources.Value));
+                }
             }
             
             // generate .config file for bridge executable
